test: KEEP-947 move the chronicle Tier 2 suite to the mainnet fork
Replaces the hard-skipped Sepolia suite (blocked since 2026-07-02 on
cold-contract fetches through the throttled public Sepolia upstream)
with chronicle/ethereum gated on ANVIL_FORK_MAINNET_URL, whitelisted by
the fork ward-kiss preflight. The Sepolia testData is removed with it:
its only consumers were the deleted suite, the Tier 1 sepolia sim
registration, and the seeder's sepolia dashboard fixtures, which the
chain-1 fixtures replace (the registry keeps every Sepolia feed address,
so the product surface is unchanged).

Verified on the local rig end to end through the production app:
12 executed / 12 passed / 0 failed (5 value-asserted reads), 5
documented skips. Chronicle is all reads, so this holds without real
Turnkey signing. Goldens: chronicle.json drops the 11155111 section.
