+++ protocols/chronicle.ts
@@ -90,71 +90,6 @@ const TEST_DATA: ProtocolTestData = {
         "SelfKisser is testnet-only; mainnet whitelisting is ward-gated (fork provisioning impersonates a ward)",
     },
   },
-  "11155111": {
-    setup: {
-      minNativeHuman: "0.01",
-      requiredTokens: [],
-      approvals: [],
-      protocolSteps: [
-        {
-          protocol: "chronicle",
-          action: "self-kiss",
-          inputs: { oracle: contract("ethUsd") },
-        },
-        {
-          protocol: "chronicle",
-          action: "self-kiss",
-          inputs: { oracle: contract("btcUsd") },
-        },
-        {
-          protocol: "chronicle",
-          action: "self-kiss",
-          inputs: { oracle: contract("daiUsd") },
-        },
-        {
-          protocol: "chronicle",
-          action: "self-kiss",
-          inputs: { oracle: contract("usdcUsd") },
-        },
-        {
-          protocol: "chronicle",
-          action: "self-kiss",
-          inputs: { oracle: contract("usdtUsd") },
-        },
-        {
-          protocol: "chronicle",
-          action: "self-kiss",
-          inputs: { oracle: contract("linkUsd") },
-        },
-      ],
-    },
-    actions: {
-      "eth-usd-read": {},
-      "eth-usd-read-with-age": {},
-      "btc-usd-read": {},
-      "btc-usd-read-with-age": {},
-      "dai-usd-read": {},
-      "dai-usd-read-with-age": {},
-      "usdc-usd-read": {},
-      "usdc-usd-read-with-age": {},
-      "usdt-usd-read": {},
-      "usdt-usd-read-with-age": {},
-      "link-usd-read": {},
-      "link-usd-read-with-age": {},
-      // customOracle is userSpecifiedAddress; point the generic read actions
-      // at the ETH/USD oracle the setup workflow already self-kissed above,
-      // so read/readWithAge (which require whitelisting) succeed too, not
-      // just tryRead/tryReadWithAge.
-      read: { contractAddress: contract("ethUsd") },
-      "try-read": { contractAddress: contract("ethUsd") },
-      "read-with-age": { contractAddress: contract("ethUsd") },
-      "try-read-with-age": { contractAddress: contract("ethUsd") },
-      "self-kiss": { oracle: contract("ethUsd") },
-    },
-    skipped: {
-      "self-kiss": "used in setup; re-running would error if already kissed",
-    },
-  },
 };
 
 // Minimal oracle ABI used for named feed contracts (read + readWithAge only).
