+++ tests/unit/__goldens__/protocol-calldata/chronicle.json
@@ -85,92 +85,5 @@
       "data": "0xc529a6790000000000000000000000000000000000000000000000000000000000000000",
       "skipped": true
     }
-  },
-  "11155111": {
-    "eth-usd-read": {
-      "to": "0xdd6D76262Fd7BdDe428dcfCd94386EbAe0151603",
-      "data": "0x57de26a4",
-      "skipped": false
-    },
-    "eth-usd-read-with-age": {
-      "to": "0xdd6D76262Fd7BdDe428dcfCd94386EbAe0151603",
-      "data": "0x393e5ede",
-      "skipped": false
-    },
-    "btc-usd-read": {
-      "to": "0x6edF073c4Bd934d3916AA6dDAC4255ccB2b7c0f0",
-      "data": "0x57de26a4",
-      "skipped": false
-    },
-    "btc-usd-read-with-age": {
-      "to": "0x6edF073c4Bd934d3916AA6dDAC4255ccB2b7c0f0",
-      "data": "0x393e5ede",
-      "skipped": false
-    },
-    "dai-usd-read": {
-      "to": "0xaf900d10f197762794C41dac395C5b8112eD13E1",
-      "data": "0x57de26a4",
-      "skipped": false
-    },
-    "dai-usd-read-with-age": {
-      "to": "0xaf900d10f197762794C41dac395C5b8112eD13E1",
-      "data": "0x393e5ede",
-      "skipped": false
-    },
-    "usdc-usd-read": {
-      "to": "0xb34d784dc8E7cD240Fe1F318e282dFdD13C389AC",
-      "data": "0x57de26a4",
-      "skipped": false
-    },
-    "usdc-usd-read-with-age": {
-      "to": "0xb34d784dc8E7cD240Fe1F318e282dFdD13C389AC",
-      "data": "0x393e5ede",
-      "skipped": false
-    },
-    "usdt-usd-read": {
-      "to": "0x8c852EEC6ae356FeDf5d7b824E254f7d94Ac6824",
-      "data": "0x57de26a4",
-      "skipped": false
-    },
-    "usdt-usd-read-with-age": {
-      "to": "0x8c852EEC6ae356FeDf5d7b824E254f7d94Ac6824",
-      "data": "0x393e5ede",
-      "skipped": false
-    },
-    "link-usd-read": {
-      "to": "0x260c182f0054BF244a8e38d7C475b6d9f67AeAc1",
-      "data": "0x57de26a4",
-      "skipped": false
-    },
-    "link-usd-read-with-age": {
-      "to": "0x260c182f0054BF244a8e38d7C475b6d9f67AeAc1",
-      "data": "0x393e5ede",
-      "skipped": false
-    },
-    "read": {
-      "to": "0xdd6D76262Fd7BdDe428dcfCd94386EbAe0151603",
-      "data": "0x57de26a4",
-      "skipped": false
-    },
-    "try-read": {
-      "to": "0xdd6D76262Fd7BdDe428dcfCd94386EbAe0151603",
-      "data": "0x10b07b71",
-      "skipped": false
-    },
-    "read-with-age": {
-      "to": "0xdd6D76262Fd7BdDe428dcfCd94386EbAe0151603",
-      "data": "0x393e5ede",
-      "skipped": false
-    },
-    "try-read-with-age": {
-      "to": "0xdd6D76262Fd7BdDe428dcfCd94386EbAe0151603",
-      "data": "0xceed3ef2",
-      "skipped": false
-    },
-    "self-kiss": {
-      "to": "0x9eE458DefDc50409DbF153515DA54Ff5B744e533",
-      "data": "0xc529a679000000000000000000000000dd6d76262fd7bdde428dcfcd94386ebae0151603",
-      "skipped": true
-    }
   }
 }
